refactor: use mixins and helpers for seguimientos
