+++ src/views/PanelSeguimientos/Seguimientos.vue
@@ -255,16 +255,15 @@ import GuiasEmpresaTab from './components/GuiasEmpresaTab.vue'
 import GuiaEmpresaModal from './components/GuiaEmpresaModal.vue'
 import SeguimientoModal from './components/SeguimientoModal.vue'
 import store from '@/store'
-import ordenes from '@/store/ordenesV3' // Módulo Vuex para órdenes
-import empresasV3 from '@/store/empresasV3' // Módulo Vuex para empresas
-import roles from '@/store/roles' // Módulo Vuex para roles de usuario
-import guias from '@/store/guias' // Importamos el módulo de guías
 import excel from 'exceljs'
 import { saveAs } from 'file-saver'
 import jsPDF from 'jspdf'
   
+import apiMixin from '@/mixins/seguimientosApiMixin'
+import { construirTimelineOrden, construirTimelineGuia } from '@/helpers/timeline'
   export default {
     name: 'SeguimientosOrdenesGuias', // Nuevo nombre para el componente
+  mixins: [apiMixin],
     components: { SelectorEmpresa, OrdenesTab, GuiasTab, GuiasEmpresaTab, SeguimientoModal, GuiaEmpresaModal },
     data() {
       return {
@@ -606,11 +605,11 @@ import jsPDF from 'jspdf'
         if (!this.modalData) return [];
         // Si el modal es de tipo 'orden', usa la función de timeline de órdenes.
         if (this.modalType === 'orden') {
-          return this._construirTimelineOrden(this.modalData);
+          return construirTimelineOrden(this.modalData);
         }
         // Si el modal es de tipo 'guia', usa la función de timeline de guías.
         else if (this.modalType === 'guia') {
-          return this._construirTimelineGuia(this.modalData);
+          return construirTimelineGuia(this.modalData);
         }
         return []; // Si no es ni orden ni guía, devuelve un array vacío.
       }
@@ -657,863 +656,6 @@ import jsPDF from 'jspdf'
         this.popularAmbasListas()
       },
   
-      /**
-       * `fetchEmpresas`:
-       * Método asíncrono para cargar la lista de empresas activas desde Vuex.
-       * Almacena el resultado en `this.listaEmpresas`.
-       */
-      async fetchEmpresas() {
-        console.log("fetchEmpresas: Iniciando carga de empresas.");
-        try {
-          // Dispara la acción Vuex para cargar la lista de empresas.
-          await store.dispatch('empresas/cargarListaEmpresas', 'SoloActivas');
-          // Asigna la lista de empresas del store localmente.
-          this.listaEmpresas = store.state.empresas.listaEmpresas || [];
-          console.log("fetchEmpresas: Empresas cargadas exitosamente:", this.listaEmpresas.length);
-        } catch (error) {
-          console.error("fetchEmpresas: Error al cargar empresas:", error);
-          this.listaEmpresas = []; // En caso de error, la lista queda vacía.
-        }
-      },
-  
-      /**
-       * `eligioEmpresa`:
-       * Método que se ejecuta cuando el usuario selecciona una empresa del `SelectorEmpresa`.
-       * Realiza la configuración inicial para la empresa seleccionada y recarga las listas.
-       * @param {number} idEmpresa - El ID de la empresa seleccionada.
-       */
-      async eligioEmpresa(idEmpresa) {
-        console.log("eligioEmpresa: Empresa seleccionada con ID:", idEmpresa);
-
-        const id = Number(idEmpresa)
-        if (isNaN(id) || id <= 0) {
-          this.mostrarError('Seleccione una empresa válida')
-          return
-        }
-        // Limpia los datos del modal y lo oculta.
-        this.modalData = null;
-        this.showModal = false;
-        this.errorAlCargar = null; // Limpia mensajes de error previos.
-
-        try {
-          // Obtiene la configuración detallada de la empresa seleccionada.
-          const resp = await empresasV3.getOneById(id);
-          console.log("eligioEmpresa: Configuración de empresa recibida:", resp);
-          // Asigna propiedades de la empresa (configuración, lote, stock).
-          this.estaEmpresa = resp;
-          this.empresaPosicionado = resp.StockPosicionado;
-          this.empresaPorLote = resp.LOTE;
-  
-          // Si es un cliente textil, obtiene las unidades totales en base.
-          if (resp.ClienteTextil === true) {
-            this.unidadesTotalesEnBase = await ordenes.contadorBultosDia(
-              idEmpresa,
-              this.getHoy()
-            );
-            console.log("eligioEmpresa: Unidades totales en base (cliente textil):", this.unidadesTotalesEnBase);
-          } else {
-            this.unidadesTotalesEnBase = 0;
-          }
-        } catch (error) {
-          console.error("eligioEmpresa: Error al configurar empresa:", error);
-          // Restablece las propiedades de la empresa en caso de error.
-          this.estaEmpresa = {};
-          this.empresaPosicionado = false;
-          this.empresaPorLote = false;
-          this.unidadesTotalesEnBase = 0;
-          this.errorAlCargar = "No se pudo cargar la configuración de la empresa.";
-        }
-  
-        // Actualiza el ID de la empresa seleccionada y reinicia las paginaciones.
-        this.idEmpresa = id;
-        this.pageOrdenes = 1;
-        this.pageGuias = 1;
-        // Llama a la función para cargar ambas listas (órdenes y guías).
-        await this.popularAmbasListas();
-      },
-  
-      /**
-       * `popularAmbasListas`:
-       * Método asíncrono que coordina la carga de órdenes y guías en paralelo.
-       * Se dispara cuando cambian la empresa o las fechas de filtro.
-       */
-      async popularAmbasListas() {
-        console.log("popularAmbasListas: Iniciando carga de datos...");
-        console.log("Parámetros actuales:", {
-          idEmpresa: this.idEmpresa,
-          fechaDesde: this.fechaDesde,
-          fechaHasta: this.fechaHasta
-        });
-        
-        // Solo procede si hay una empresa y fechas de filtro válidas.
-        if (this.idEmpresa > 0 && this.fechaDesde && this.fechaHasta) {
-          console.log("popularAmbasListas: Parámetros válidos, iniciando carga...");
-          this.loading = true; // Activa el spinner de carga general.
-          this.errorAlCargar = null; // Limpia errores previos.
-          
-          try {
-            console.log("popularAmbasListas: Iniciando carga en paralelo de órdenes y guías...");
-            
-            // Ejecuta ambas funciones de carga en paralelo usando `Promise.all`.
-            await Promise.all([
-              this.popularListaDeOrdenes().catch(error => {
-                console.error("Error en popularListaDeOrdenes:", error);
-                throw error; // Relanza el error para manejarlo en el catch externo
-              }),
-              this.popularListaDeGuias().catch(error => {
-                console.error("Error en popularListaDeGuias:", error);
-                throw error; // Relanza el error para manejarlo en el catch externo
-              }),
-              this.popularListaGuiasEmpresa().catch(error => {
-                console.error('Error en popularListaGuiasEmpresa:', error)
-                throw error
-              })
-            ]);
-            
-            console.log("popularAmbasListas: Carga de órdenes y guías completada exitosamente.");
-          } catch (error) {
-            console.error("popularAmbasListas: Error al cargar datos:", {
-              error: error,
-              message: error.message,
-              stack: error.stack,
-              response: error.response?.data || 'No hay datos de respuesta'
-            });
-            
-            // Muestra un mensaje de error más descriptivo
-            const errorMessage = error.response?.data?.message || error.message || "Ocurrió un error al cargar los datos. Intente nuevamente.";
-            this.errorAlCargar = errorMessage;
-            
-            // Si es un error de autenticación, redirigir al login
-            if (error.response?.status === 401) {
-              console.warn("Sesión expirada, redirigiendo a login...");
-              this.$router.push('/login');
-            }
-            
-            // Limpia las listas para asegurar consistencia
-            this.todasLasOrdenes = [];
-            this.todasLasGuias = [];
-          } finally {
-            console.log("popularAmbasListas: Finalizando carga, desactivando spinner...");
-            this.loading = false; // Desactiva el spinner de carga.
-          }
-        } else {
-          console.log("popularAmbasListas: No hay empresa o fechas seleccionadas, limpiando listas.", {
-            tieneEmpresa: this.idEmpresa > 0,
-            tieneFechaDesde: !!this.fechaDesde,
-            tieneFechaHasta: !!this.fechaHasta
-          });
-          
-          // Si no hay filtros válidos, limpia las listas y los estados de carga/error.
-          this.todasLasOrdenes = [];
-          this.todasLasGuias = [];
-          this.errorAlCargar = null;
-          this.loading = false;
-        }
-      },
-  
-      /**
-       * `popularListaDeOrdenes`:
-       * Método asíncrono para cargar y procesar la lista de órdenes usando el nuevo endpoint.
-       * Filtra las órdenes por empresa y rango de fechas, las formatea y las ordena.
-       */
-      async popularListaDeOrdenes() {
-        console.log("popularListaDeOrdenes: Iniciando carga de órdenes con el nuevo endpoint.");
-        console.log("Parámetros actuales:", {
-          idEmpresa: this.idEmpresa,
-          fechaDesde: this.fechaDesde,
-          fechaHasta: this.fechaHasta,
-          usuarioActual: store.state.usuarios.usuarioActual
-        });
-        
-        this.todasLasOrdenes = []; // Limpia la lista de órdenes antes de la carga.
-  
-        // Bloque para verificar roles de usuario (se mantiene del código original).
-        try {
-          const userId = store.state.usuarios.usuarioActual?.Id;
-          console.log("popularListaDeOrdenes: Obteniendo roles para userId:", userId);
-          
-          if (!userId) {
-            console.error("No se pudo obtener el ID del usuario actual");
-            throw new Error('No se pudo identificar al usuario actual');
-          }
-          
-          const rolesUser = await roles.getUserRolesById(userId);
-          console.log("Roles del usuario:", rolesUser);
-          
-          this.rolPermitido =
-            !store.state.usuarios.usuarioActual?.IdEmpresa &&
-            rolesUser?.[0]?.IdRole === 1;
-            
-          console.log("popularListaDeOrdenes: rolPermitido:", this.rolPermitido);
-        } catch (error) {
-          console.error("popularListaDeOrdenes: Error al verificar roles de usuario:", error);
-          this.rolPermitido = false;
-        }
-  
-        try {
-          console.log(`popularListaDeOrdenes: Llamando a ordenes.getOrdenesByEmpresaPeriodoConDestinos para idEmpresa: ${this.idEmpresa}, fechas: ${this.fechaDesde} a ${this.fechaHasta}`);
-          
-          console.log("Llamando a getOrdenesByEmpresaPeriodoConDestinos con:", {
-            idEmpresa: this.idEmpresa,
-            fechaDesde: this.fechaDesde,
-            fechaHasta: this.fechaHasta
-          });
-          
-          // Llama al nuevo método del servicio de órdenes
-          let response;
-          try {
-            console.log(`Solicitando órdenes para empresa ${this.idEmpresa} desde ${this.fechaDesde} hasta ${this.fechaHasta}`);
-            response = await ordenes.getOrdenesByEmpresaPeriodoConDestinos(this.idEmpresa, this.fechaDesde, this.fechaHasta);
-            
-            // Depuración profunda de la respuesta
-            console.log("Respuesta recibida:", response);
-            console.log("Tipo de respuesta:", typeof response);
-            console.log("Es array?", Array.isArray(response));
-            console.log("Tiene propiedad status?", response && 'status' in response);
-            console.log("Tiene propiedad data?", response && 'data' in response);
-            console.log("Data es array?", response && response.data && Array.isArray(response.data));
-            
-          } catch (apiError) {
-            console.error("Error en la llamada a getOrdenesByEmpresaPeriodoConDestinos:", {
-              error: apiError,
-              message: apiError.message,
-              stack: apiError.stack,
-              response: apiError.response?.data || 'No hay datos de respuesta'
-            });
-            throw new Error('Error al conectar con el servidor. Verifica tu conexión e intenta nuevamente.');
-          }
-  
-          let data = [];
-          
-          // Manejo de la respuesta según el formato esperado
-          if (response && response.status === 'OK' && Array.isArray(response.data)) {
-            // Formato esperado: {status: 'OK', data: [...]}
-            console.log(`Se recibieron ${response.data.length} órdenes del servidor`);
-            data = response.data;
-            
-            // Verificar si hay errores en la respuesta
-            if (response.errors && response.errors.length > 0) {
-              console.warn('La respuesta contiene errores:', response.errors);
-            }
-          } else if (response && Array.isArray(response)) {
-            // Si por alguna razón la respuesta es directamente un array
-            console.warn('La respuesta es un array directo, no el formato esperado. Total de órdenes:', response.length);
-            data = response;
-          } else if (response && response.data === undefined) {
-            // Si la respuesta no tiene data, asumimos array vacío
-            console.log("La respuesta no contiene datos (array vacío)");
-            data = [];
-          } else {
-            // Formato inesperado
-            console.error("Formato de respuesta inesperado:", response);
-            // Lanzamos un error más descriptivo
-            throw new Error('Formato de respuesta inesperado del servidor. Contacte a soporte.');
-          }
-          
-          console.log(`Total de órdenes a procesar: ${data.length}`);
-  
-          // Agrupar los datos por IdOrden y consolidar los productos
-          const ordenesAgrupadas = {};
-          data.forEach(item => {
-            if (!ordenesAgrupadas[item.IdOrden]) {
-              ordenesAgrupadas[item.IdOrden] = {
-                IdOrden: item.IdOrden,
-                Numero: item.Numero,
-                Fecha: item.Fecha, // Fecha de la orden (que ahora es la fecha de distribución/retiro)
-                Estado: item.Estado, // Estado numérico
-                Tipo: item.Tipo,
-                preOrden: item.preOrden,
-                retira_cliente: item.retira_cliente,
-                Valor: item.Valor,
-                Observaciones: item.Observaciones,
-                id_integracion: item.id_integracion,
-                IdDestino: item.IdDestino,
-                NombreDestino: item.NombreDestino,
-                DomicilioDestino: item.DomicilioDestino,
-                LocalidadDestino: item.LocalidadDestino,
-                CodigoPostalDestino: item.CodigoPostalDestino,
-                ObservacionesDestino: item.ObservacionesDestino,
-                IdEmpresaDestino: item.IdEmpresaDestino,
-                // Propiedades de la empresa de la orden (si vinieran por este endpoint, si no, se usan las del componente)
-                // Aquí asumo que la empresa de la orden es this.estaEmpresa
-                NombreEmpresa: this.estaEmpresa.RazonSocial || 'N/A', 
-                // Para la guía asociada, necesitamos buscarla o determinarla si viene en el JSON
-                IdGuia: item.IdGuia || -1, // Asumo que el IdGuia viene en el item
-                productos: [] // Inicializamos el array de productos
-              };
-            }
-            // Agregamos el producto a la orden correspondiente
-            ordenesAgrupadas[item.IdOrden].productos.push({
-              IdDetalle: item.IdDetalle,
-              Unidades: item.Unidades,
-              Precio: item.Precio,
-              loteCompleto: item.loteCompleto,
-              IdProducto: item.IdProducto,
-              Barcode: item.Barcode,
-              NombreProducto: item.NombreProducto,
-              CodeEmpresa: item.CodeEmpresa,
-              Alto: item.Alto,
-              Ancho: item.Ancho,
-              Largo: item.Largo,
-              Peso: item.Peso
-              // Si hay campos como 'lote' o 'Partida' en el JSON original para los productos, agrégalos aquí.
-              // Por ejemplo: lote: item.lote, Partida: item.Partida
-            });
-          });
-  
-          let todasOrdenes = Object.values(ordenesAgrupadas);
-          console.log("popularListaDeOrdenes: Órdenes agrupadas y productos consolidados:", todasOrdenes.length);
-
-          // Obtener historial de cada orden y formatear datos para la tabla
-          for (const o of todasOrdenes) {
-            try {
-              const historico = await ordenes.getHistoricoEstados(o.IdOrden)
-              const historialOrden = Array.isArray(historico) ? historico : []
-              historialOrden.sort((a, b) => new Date(a.Fecha) - new Date(b.Fecha))
-              o.historialEstados = historialOrden
-
-              const ultimo = historialOrden[historialOrden.length - 1]
-              if (ultimo) {
-                o.Estado = ultimo.Estado
-              }
-            } catch (histErr) {
-              console.error(`popularListaDeOrdenes: Error al obtener historial para orden ${o.IdOrden}:`, histErr)
-              o.historialEstados = []
-            }
-
-            // La 'Fecha' que viene es la Fecha de distribución/retiro.
-            // Para la "Fecha Creación" en la tabla, usaremos la "Fecha" del JSON que representa la Fecha de la orden
-            // (que en este nuevo endpoint es la fecha de distribución/retiro).
-            // Si necesitas la fecha de AltaRegistro para la columna "Fecha Creación", cámbialo aquí.
-            o.Fecha = o.Fecha ? new Date(o.Fecha).toLocaleDateString('es-AR') : 'N/A'
-            // o.AltaRegistro = o.AltaRegistro ? new Date(o.AltaRegistro).toLocaleDateString('es-AR') : 'N/A'; // Si quieres usar AltaRegistro
-
-            // Traduce el estado numérico de la orden a un estado textual legible.
-            switch (o.Estado) {
-              case 1: o.NombreEstado = 'Pendiente'; break;
-              case 2: o.NombreEstado = 'Preparado'; break;
-              case 3: o.NombreEstado = 'A distribución'; break; // Estado textual que indica que tiene guía
-              case 4: o.NombreEstado = 'Anulado'; break;
-              case 5: o.NombreEstado = 'Retira Cliente'; break;
-              default: o.NombreEstado = `Desconocido (${o.Estado})`;
-            }
-
-            // Asigna el estado textual al campo `Estado` que la tabla usa para la columna.
-            o.Estado = o.NombreEstado // Para la tabla
-
-            o.tipo = o.preOrden === true || o.preOrden === 1 ? 'Pre-Orden' : 'Orden'
-            o.preOrdenDisplay = o.tipo
-
-            // Aseguramos que IdGuia esté presente, aunque puede ser -1 si no hay guía
-            o.IdGuia = o.IdGuia || -1
-          }
-  
-          // Ordena las órdenes por fecha de la orden de forma descendente (más recientes primero).
-          // Usamos la propiedad 'Fecha' que ya ha sido procesada y formateada como fecha de la orden.
-          todasOrdenes.sort(
-            (a, b) => {
-              const dateA = new Date(a.Fecha.split('/').reverse().join('-')); // Convertir a formato YYYY-MM-DD para Date
-              const dateB = new Date(b.Fecha.split('/').reverse().join('-'));
-              return dateB.getTime() - dateA.getTime();
-            }
-          );
-  
-          this.todasLasOrdenes = todasOrdenes;
-          console.log("popularListaDeOrdenes: Total de órdenes procesadas y listas para la tabla:", this.todasLasOrdenes.length);
-        } catch (error) {
-          console.error("popularListaDeOrdenes: Error al cargar órdenes:", error);
-          throw new Error('No se pudieron cargar las órdenes. Intente nuevamente.');
-        }
-      },
-  
-      /**
-       * `popularListaDeGuias`:
-       * Método asíncrono para cargar y procesar la lista de guías.
-       * Utiliza el endpoint `getByPeriodoIdEmpresa` para obtener las guías.
-       */
-      async popularListaDeGuias() {
-        console.log("popularListaDeGuias: Iniciando carga de guías.");
-        this.todasLasGuias = []; // Limpia la lista de guías antes de la carga.
-        try {
-          // Llama al servicio de guías para obtenerlas por periodo y ID de empresa.
-          console.log(`popularListaDeGuias: Llamando a guias.getByPeriodoIdEmpresa con fechaDesde: ${this.fechaDesde}, fechaHasta: ${this.fechaHasta}, idEmpresa: ${this.idEmpresa}`);
-          const response = await guias.getByPeriodoIdEmpresa(this.fechaDesde, this.fechaHasta, this.idEmpresa);
-          console.log("popularListaDeGuias: Respuesta de guias.getByPeriodoIdEmpresa recibida:", response);
-  
-          let todas = [];
-          // Verifica el formato de la respuesta para obtener los datos.
-          if (response && response.status === 'OK' && Array.isArray(response.data)) {
-            todas = response.data;
-          } else if (Array.isArray(response)) {
-            todas = response; // Si la respuesta ya es el array directamente.
-          } else {
-            // Lanza un error si el formato es inesperado.
-            throw new Error('Formato de guías inesperado. Contacte a soporte.');
-          }
-  
-          // Formatea los datos de cada guía para su visualización en la tabla.
-          todas.forEach((g) => {
-            g.FechaOriginalDate = new Date(g.FechaOriginal); // Crea un objeto Date para facilitar filtros y ordenamiento.
-            g.FechaOriginal = g.FechaOriginal ? new Date(g.FechaOriginal).toLocaleDateString('es-AR') : 'N/A'; // Formatea la fecha original.
-            g.NombreCliente = g.NombreCliente || 'N/A';
-            g.NombreDestino = g.NombreDestino || 'N/A';
-            g.Remitos = g.Remitos || 'N/A';
-            g.Estado = g.Estado || 'Desconocido'; // Asegura que el estado exista y sea textual.
-          });
-  
-          // Ordena las guías por fecha de creación de forma descendente.
-          todas.sort((a, b) => b.FechaOriginalDate.getTime() - a.FechaOriginalDate.getTime());
-          this.todasLasGuias = todas;
-          console.log("popularListaDeGuias: Total de guías procesadas:", this.todasLasGuias.length);
-        } catch (error) {
-          console.error("popularListaDeGuias: Error al cargar guías:", error);
-          // Lanza el error para que `popularAmbasListas` lo capture.
-          throw new Error('No se pudieron cargar las guías. Intente nuevamente.');
-        }
-      },
-
-      async popularListaGuiasEmpresa() {
-        console.log('popularListaGuiasEmpresa: Iniciando carga')
-        this.todasLasGuiasEmpresa = []
-        try {
-          const response = await guias.getByPeriodoIdEmpresa(this.fechaDesde, this.fechaHasta, this.idEmpresa)
-          let todas = []
-          if (response && response.status === 'OK' && Array.isArray(response.data)) {
-            todas = response.data
-          } else if (Array.isArray(response)) {
-            todas = response
-          } else {
-            throw new Error('Formato de guías inesperado')
-          }
-
-          todas.forEach(g => {
-            g.FechaOriginalDate = new Date(g.FechaOriginal)
-            g.FechaOriginal = g.FechaOriginal ? new Date(g.FechaOriginal).toLocaleDateString('es-AR') : 'N/A'
-            g.FechaEntrega = g.FechaEntrega ? new Date(g.FechaEntrega).toLocaleDateString('es-AR') : 'N/A'
-            g.Bultos = g.Bultos || 0
-            g.Remitos = g.Remitos || 'N/A'
-            g.Estado = g.Estado || 'Desconocido'
-          })
-
-          todas.sort((a, b) => b.FechaOriginalDate.getTime() - a.FechaOriginalDate.getTime())
-          this.todasLasGuiasEmpresa = todas
-          console.log('popularListaGuiasEmpresa: Total', this.todasLasGuiasEmpresa.length)
-        } catch (error) {
-          console.error('popularListaGuiasEmpresa: Error', error)
-          throw new Error('No se pudieron cargar las guías por empresa')
-        }
-      },
-  
-      /**
-       * `verGuiaAsociada`:
-       * Método que se dispara al hacer clic en el chip "Seguimiento aquí" de una orden.
-       * Cambia a la pestaña de guías y abre el modal de detalle de la guía asociada.
-       * @param {object} ordenItem - El objeto completo de la orden desde la tabla.
-       */
-      async verGuiaAsociada(ordenItem) {
-        console.log("verGuiaAsociada: Intentando ver guía asociada para orden:", ordenItem.Numero, "IdGuia:", ordenItem.IdGuia);
-        // Asegura que hay un IdGuia válido.
-        if (ordenItem.IdGuia > 0) {
-          // Busca la guía correspondiente en la lista de guías ya cargadas.
-          const guiaAsociada = this.todasLasGuias.find(g => g.Id === ordenItem.IdGuia);
-  
-          if (guiaAsociada) {
-            console.log("verGuiaAsociada: Guía asociada encontrada localmente:", guiaAsociada.Comprobante);
-            this.tab = 'tab-guias'; // Cambia a la pestaña de Guías.
-            // Abre el modal de la guía, pasando el objeto completo de la guía.
-            await this.openModal('guia', guiaAsociada);
-          } else {
-            // Si por alguna razón la guía no está en `todasLasGuias` (ej. fue cargada fuera del rango de fechas de guías),
-            // se intenta obtenerla directamente de la API.
-            console.warn("verGuiaAsociada: Guía asociada no encontrada en la lista local. Intentando obtener de la API.");
-            try {
-              const response = await guias.getById(ordenItem.IdGuia);
-              const retrievedGuia = (response && response.data) ? response.data : (response && response.Id ? response : null);
-              if (retrievedGuia) {
-                console.log("verGuiaAsociada: Guía obtenida de la API:", retrievedGuia.Comprobante);
-                // Si se obtuvo, la agrega a la lista de guías (opcional, para futuras interacciones)
-                // y abre el modal.
-                // Formatea la fecha de la guía obtenida directamente para que sea consistente con la lista.
-                retrievedGuia.FechaOriginal = retrievedGuia.FechaOriginal ? new Date(retrievedGuia.FechaOriginal).toLocaleDateString('es-AR') : 'N/A';
-                retrievedGuia.Estado = retrievedGuia.Estado || 'Desconocido';
-                this.todasLasGuias.push(retrievedGuia); // Añade la guía a la lista para que se muestre en la tabla si corresponde
-                this.tab = 'tab-guias';
-                await this.openModal('guia', retrievedGuia);
-              } else {
-                this.errorAlCargar = "No se pudo encontrar la guía asociada.";
-                console.error("verGuiaAsociada: No se pudo obtener la guía de la API.");
-              }
-            } catch (error) {
-              this.errorAlCargar = `Error al cargar la guía asociada: ${error.message || error}`;
-              console.error("verGuiaAsociada: Error al intentar obtener guía de la API:", error);
-            }
-          }
-        } else {
-          console.warn("verGuiaAsociada: La orden no tiene un IdGuia válido.");
-          this.errorAlCargar = "Esta orden no tiene una guía asociada para seguimiento.";
-        }
-      },
-  
-      /**
-       * `openModal`:
-       * Método asíncrono que se dispara al hacer clic en el ícono "ojo" de una fila.
-       * Carga los detalles del elemento (orden o guía) y los muestra en un modal.
-       * @param {string} type - 'orden' o 'guia', indica el tipo de elemento a abrir.
-       * @param {number|object} item - Para 'orden', es el ID de la orden. Para 'guia', es el objeto completo de la guía.
-       */
-      async openModal(type, item) {
-        console.log(`openModal: Abriendo modal para tipo: ${type}, item:`, item);
-        this.modalData = null; // Limpia datos de modal previos.
-        this.loading = true; // Activa el spinner dentro del modal.
-        this.errorAlCargar = null; // Limpia errores previos.
-        this.modalType = type; // Establece el tipo del modal (orden/guía).
-  
-        try {
-          let dataToModal;
-          if (type === 'orden') {
-            // Si es una orden, se espera que `item` sea el ID de la orden.
-            // Buscamos la orden directamente de `todasLasOrdenes` que ya contiene todos los detalles.
-            const found = this.todasLasOrdenes.find(o => o.IdOrden === item);
-            dataToModal = found ? JSON.parse(JSON.stringify(found)) : null;
-
-            if (dataToModal) {
-              // Obtener historial de estados para la orden
-              const historico = await ordenes.getHistoricoEstados(item)
-              const historialOrden = Array.isArray(historico) ? historico : []
-              historialOrden.sort((a, b) => new Date(a.Fecha) - new Date(b.Fecha))
-
-              dataToModal.historialEstados = historialOrden
-
-              const ultimo = historialOrden[historialOrden.length - 1]
-              if (ultimo) {
-                dataToModal.Estado = ultimo.Estado
-                switch (ultimo.Estado) {
-                  case 1: dataToModal.NombreEstado = 'Pendiente'; break;
-                  case 2: dataToModal.NombreEstado = 'Preparado'; break;
-                  case 3: dataToModal.NombreEstado = 'A distribución'; break;
-                  case 4: dataToModal.NombreEstado = 'Anulado'; break;
-                  case 5: dataToModal.NombreEstado = 'Retira Cliente'; break;
-                  default: dataToModal.NombreEstado = `Desconocido (${ultimo.Estado})`;
-                }
-              } else {
-                switch (dataToModal.Estado) {
-                  case 1: dataToModal.NombreEstado = 'Pendiente'; break;
-                  case 2: dataToModal.NombreEstado = 'Preparado'; break;
-                  case 3: dataToModal.NombreEstado = 'A distribución'; break;
-                  case 4: dataToModal.NombreEstado = 'Anulado'; break;
-                  case 5: dataToModal.NombreEstado = 'Retira Cliente'; break;
-                  default: dataToModal.NombreEstado = `Desconocido (${dataToModal.Estado})`;
-                }
-              }
-
-              dataToModal.Fecha = dataToModal.Fecha ? new Date(dataToModal.Fecha).toLocaleDateString('es-AR') : 'N/A'
-              dataToModal.NombreEmpresa = dataToModal.NombreEmpresa || this.estaEmpresa.RazonSocial || 'N/A'
-              console.log('openModal: Datos de orden para modal procesados:', dataToModal)
-            } else {
-              throw new Error('No se encontraron detalles válidos para esta orden.');
-            }
-          } else if (type === 'guia') {
-            // Para las guías, obtenemos el detalle completo desde la API utilizando su Id.
-            console.log("openModal: Procesando datos de guía para modal:", item);
-
-            const response = await guias.getById(item.Id);
-            const guiaCompleta = (response && response.data)
-              ? response.data
-              : (response && response.Id ? response : null);
-
-            if (!guiaCompleta) {
-              throw new Error('No se encontraron detalles válidos para esta guía.');
-            }
-
-            dataToModal = { ...item, ...guiaCompleta };
-            dataToModal.productos = Array.isArray(guiaCompleta.productos) ? guiaCompleta.productos : [];
-
-            // Formatea fechas específicas de la guía para el modal.
-            dataToModal.FechaOriginal = dataToModal.FechaOriginal ? new Date(dataToModal.FechaOriginal).toLocaleDateString('es-AR') : 'N/A';
-            // Para la fecha de no entrega, se usa la `Fecha` de la guía si el estado es `NO ENTREGADO`.
-            if (dataToModal.Estado === 'NO ENTREGADO' && dataToModal.Fecha) {
-              dataToModal.FechaNoEntregado = new Date(dataToModal.Fecha).toLocaleDateString('es-AR');
-            } else {
-              dataToModal.FechaNoEntregado = 'N/A';
-            }
-
-            console.log("openModal: Datos de guía para modal procesados:", dataToModal);
-          }
-  
-          // Si se obtuvieron datos válidos para el modal, los asigna y lo muestra.
-          if (dataToModal && (dataToModal.IdOrden || dataToModal.Id)) { // IdOrden para ordenes, Id para guias
-            this.modalData = dataToModal;
-            this.showModal = true;
-            console.log("openModal: Modal abierto con datos.");
-          } else {
-            // Si `dataToModal` es nulo o incompleto después de la lógica, se lanza un error.
-            throw new Error('No se encontraron detalles o la respuesta fue inválida.');
-          }
-        } catch (error) {
-          console.error("openModal: Error al cargar detalles del modal:", error);
-          this.errorAlCargar = error.message || 'Error al cargar los detalles. Intente nuevamente.';
-        } finally {
-          this.loading = false; // Desactiva el spinner del modal.
-          console.log("openModal: Loading finalizado.");
-        }
-      },
-
-      openGuiaEmpresaModal(item) {
-        this.guiaEmpresaModalData = item
-        this.showGuiaEmpresaModal = true
-      },
-
-      closeGuiaEmpresaModal() {
-        this.guiaEmpresaModalData = null
-        this.showGuiaEmpresaModal = false
-      },
-  
-      /**
-       * `_obtenerDetalleProductos`: (Método obsoleto para el nuevo endpoint, se mantiene para contexto)
-       * Este método ya no es necesario para las órdenes si la API `getOrdenesByEmpresaPeriodoConDestinos`
-       * ya devuelve los detalles de los productos anidados.
-       */
-      async _obtenerDetalleProductos(orden) {
-        console.warn("_obtenerDetalleProductos: Este método ya no debería ser necesario si la API devuelve los productos directamente en la orden.");
-        // Se podría retornar orden.productos si existiera en la respuesta original del getById o dejarlo vacío.
-        return orden.productos || []; 
-      },
-  
-      /**
-       * `_construirTimelineOrden`:
-       * Método auxiliar para construir la línea de tiempo de una ORDEN.
-       * Define los pasos del proceso, sus iconos, fechas y clases de estado
-       * (`completed`, `current`, `current-bad`, `pending`) según el estado actual de la orden.
-       * Muestra todos los estados posibles, pintando los completados y el actual.
-       * @param {object} orden - El objeto de la orden para construir su línea de tiempo.
-       * @returns {Array} Array de objetos que representan los pasos de la línea de tiempo.
-       */
-      _construirTimelineOrden(orden) {
-        console.log('_construirTimelineOrden: Construyendo timeline para orden:', orden.Numero)
-        const historial = Array.isArray(orden.historialEstados) ? orden.historialEstados : []
-        if (!historial.length) return []
-
-        const labelMap = { 1: 'Pendiente', 2: 'Preparado', 3: 'A distribución', 4: 'Anulado', 5: 'Retira Cliente' }
-        const iconMap = {
-          1: 'mdi-file-document-edit-outline',
-          2: 'mdi-package-variant-closed-check',
-          3: 'mdi-truck-fast-outline',
-          4: 'mdi-cancel',
-          5: 'mdi-account-check-outline'
-        }
-        const colorMap = { 1: 'pendiente', 2: 'preparada', 3: 'despachada', 4: 'despachada', 5: 'despachada' }
-
-        const steps = historial.map((h, idx) => {
-          const fecha = h.Fecha ? new Date(h.Fecha).toLocaleDateString('es-AR') : 'N/A'
-          const nombre = labelMap[h.Estado] || `Estado ${h.Estado}`
-          const descripcion = h.Usuario ? `Por ${h.Usuario}` : ''
-          const statusClass = idx === historial.length - 1 ? (h.Estado === 4 ? 'current-bad' : 'current') : 'completed'
-          return {
-            id: `estado_${idx}`,
-            nombre,
-            icon: iconMap[h.Estado] || 'mdi-alert-circle-outline',
-            fecha,
-            descripcion,
-            statusClass,
-            colorClass: colorMap[h.Estado] || ''
-          }
-        })
-
-        console.log('_construirTimelineOrden: Timeline construido:', steps)
-        return steps
-      },
-  
-      /**
-       * `_construirTimelineGuia`:
-       * Método auxiliar para construir la línea de tiempo de una GUÍA.
-       * Adapta la lógica de `VistaDeTracking.vue` para definir los pasos del proceso,
-       * iconos, fechas y clases de estado según el estado actual de la guía.
-       * Muestra todos los estados posibles, pintando los completados y el actual.
-       * @param {object} guia - El objeto de la guía para construir su línea de tiempo.
-       * @returns {Array} Array de objetos que representan los pasos de la línea de tiempo.
-       */
-      _construirTimelineGuia(guia) {
-        console.log("_construirTimelineGuia: Construyendo timeline para guía:", guia.Comprobante, "Estado:", guia.Estado);
-        const timelineSteps = [];
-        const currentStatusText = guia.Estado; // El estado ya está en formato textual.
-        // Determina si la guía es para retiro en sucursal (basado en la dirección "lagos garcia 4470").
-        const isRetiraClienteSucursal = guia.Domicilio && guia.Domicilio.toLowerCase().includes("lagos garcia 4470");
-  
-        // Fechas ya formateadas en el `openModal`.
-        const fechaOriginal = guia.FechaOriginal || 'N/A';
-        const fechaNoEntregado = guia.FechaNoEntregado || 'N/A'; // Fecha específica para estado "No entregado".
-  
-        // Pasos base de la línea de tiempo con un `baseOrder` para facilitar la lógica de avance.
-        const baseSteps = [
-          {
-            id: 'preparacion',
-            nombre: 'Pedido en preparación',
-            icon: 'mdi-clipboard-list-outline',
-            fecha: fechaOriginal,
-            descripcion: 'Actualmente estamos trabajando en procesar tu pedido para que llegue a tus manos lo mas pronto posible.',
-            baseOrder: 1,
-          },
-          {
-            id: 'preparado',
-            nombre: 'Pedido preparado',
-            icon: 'mdi-cube-outline',
-            fecha: fechaOriginal,
-            descripcion: 'Tu pedido ha sido preparado y está listo para el proceso de distribución.',
-            baseOrder: 2,
-          },
-          {
-            id: 'en_cd',
-            nombre: 'En CD',
-            icon: 'mdi-warehouse',
-            fecha: fechaOriginal, // Podría ser la fecha de "En Planchada" si existe
-            descripcion: 'Tu pedido se encuentra en proceso de Cross Docking en nuestras instalaciones, ¡Pronto esta en camino hacia su destino final!',
-            baseOrder: 3,
-          },
-          {
-            id: 'en_distribucion',
-            nombre: 'En distribución',
-            icon: 'mdi-truck-fast-outline',
-            // La fecha de distribución podría ser la fecha original si está despachado, o "N/A".
-            fecha: guia.Estado === 'DESPACHADO' ? fechaOriginal : 'N/A', // Usamos la fecha original si está despachado, o "N/A"
-            descripcion: '¡Tu pedido ya está en camino!',
-            baseOrder: 4,
-          },
-        ];
-  
-        // Manejo especial para el estado "ANULADO": solo muestra un paso de anulación.
-        if (currentStatusText === 'ANULADO') {
-          return [{
-            id: 'anulado',
-            nombre: 'Guía Anulada',
-            icon: 'mdi-cancel',
-            fecha: fechaOriginal,
-            descripcion: `La guía ha sido cancelada.`,
-            statusClass: 'current-bad'
-          }];
-        }
-  
-        // Lógica para guías con retiro en sucursal ("Lagos Garcia").
-        if (isRetiraClienteSucursal) {
-          // En este caso, los primeros pasos siempre se consideran completados si la guía existe.
-          baseSteps.forEach(step => {
-            if (step.baseOrder <= 2) { // Preparación y Preparado siempre son completados para RetiraCliente
-              step.statusClass = 'completed';
-            } else {
-              step.statusClass = 'pending'; // Otros pasos como En CD y En Distribución no aplican o son pendientes
-            }
-            timelineSteps.push(step);
-          });
-  
-          // Agrega el paso final "Pedido retirado".
-          timelineSteps.push({
-            id: 'retirado',
-            nombre: 'Pedido retirado',
-            icon: 'mdi-kabaddi',
-            fecha: currentStatusText === 'ENTREGADO' ? fechaOriginal : 'N/A', // Asume fecha original si fue entregado.
-            descripcion: `El pedido fue retirado de nuestro centro de distribución.`,
-            // El estado es 'current' if the guide is 'ENTREGADO', otherwise 'pending'.
-            statusClass: currentStatusText === 'ENTREGADO' ? 'current' : 'pending'
-          });
-  
-        } else {
-          // Lógica para guías con distribución normal (no retiro en sucursal).
-  
-          // Mapeo de estados textuales de la guía a IDs de pasos del timeline para determinar el paso actual.
-          let currentStepId = '';
-          switch (currentStatusText) {
-            case 'Pedido en preparación': currentStepId = 'preparacion'; break;
-            case 'Pedido preparado': currentStepId = 'preparado'; break;
-            case 'En Planchada':
-            case 'En CD': currentStepId = 'en_cd'; break;
-            case 'Ruteado':
-            case 'DESPACHADO':
-            case 'En distribución': currentStepId = 'en_distribucion'; break;
-            case 'ENTREGADO': currentStepId = 'entregado'; break; // Un paso final especial
-            case 'PARCIAL': currentStepId = 'parcial'; break; // Un paso final especial
-            case 'NO ENTREGADO': currentStepId = 'no_entregado'; break; // Un paso final especial
-            default: currentStepId = ''; // Estado no reconocido.
-          }
-  
-          // Itera sobre los pasos base y asigna su clase de estado.
-          // Los pasos anteriores al `currentStepId` se marcan como 'completed'.
-          // El `currentStepId` se marca como 'current' o 'current-bad'.
-          // Los pasos posteriores se marcan como 'pending'.
-          baseSteps.forEach(step => {
-            let stepStatus = 'pending';
-            const currentStepIndex = baseSteps.findIndex(s => s.id === currentStepId);
-            const stepIndex = baseSteps.findIndex(s => s.id === step.id);
-  
-            if (step.id === currentStepId) { // Este es el paso actual
-              stepStatus = (currentStatusText === 'PARCIAL' || currentStatusText === 'NO ENTREGADO') ? 'current-bad' : 'current';
-            } else if (currentStepIndex !== -1 && stepIndex < currentStepIndex) { // Este paso está antes del actual
-              stepStatus = 'completed';
-            }
-            step.statusClass = stepStatus;
-            timelineSteps.push(step);
-          });
-  
-          // Agrega los pasos finales (Entregado, Parcial, No Entregado) condicionalmente.
-          // Estos pasos se añaden después de los `baseSteps` y solo si la guía ha llegado a ese estado.
-          if (currentStatusText === 'ENTREGADO') {
-              timelineSteps.push({
-                  id: 'entregado',
-                  nombre: 'Entregado',
-                  icon: 'mdi-truck-check-outline',
-                  fecha: fechaOriginal, // Asume la fecha original como fecha de entrega.
-                  descripcion: 'Tu pedido fue entrega con exito.',
-                  statusClass: 'current' // Es el último paso y el estado actual.
-              });
-          } else if (currentStatusText === 'PARCIAL') {
-              timelineSteps.push({
-                  id: 'parcial',
-                  nombre: 'Entrega parcial',
-                  icon: 'mdi-truck',
-                  fecha: fechaOriginal, // Podría ser otra fecha real del DALC.
-                  descripcion: 'Por algún motivo no se pudo entregar tu pedido de forma completa.',
-                  statusClass: 'current-bad'
-              });
-          } else if (currentStatusText === 'NO ENTREGADO') {
-              timelineSteps.push({
-                  id: 'no_entregado',
-                  nombre: 'No entregado',
-                  icon: 'mdi-truck',
-                  fecha: fechaNoEntregado,
-                  descripcion: `Al no haberse podido entregar tu pedido se recordinara.`,
-                  statusClass: 'current-bad'
-              });
-          }
-        }
-  
-        // Re-evaluación final de las clases de estado para asegurar la secuencia 'completed' -> 'current' -> 'pending'.
-        // Esta iteración final es crucial para asegurar que la línea y los iconos de los pasos previos
-        // al estado 'current' o 'current-bad' se marquen correctamente como 'completed'.
-        let foundCurrent = false;
-        for (let i = timelineSteps.length - 1; i >= 0; i--) {
-          const step = timelineSteps[i];
-          if (step.statusClass === 'current' || step.statusClass === 'current-bad') {
-            foundCurrent = true;
-          } else if (foundCurrent && step.statusClass === 'pending') {
-            // Si ya encontramos el estado actual y este paso era "pending", se marca como "completed".
-            step.statusClass = 'completed';
-          }
-        }
-
-        const colorMap = {
-          preparacion: 'pendiente',
-          preparado: 'preparada',
-          en_cd: 'despachada',
-          en_distribucion: 'despachada',
-          retirado: 'despachada',
-          entregado: 'entregado',
-          parcial: 'despachada',
-          no_entregado: 'despachada'
-        };
-        timelineSteps.forEach(step => {
-          step.colorClass = colorMap[step.id] || '';
-        });
-        console.log("_construirTimelineGuia: Timeline construido:", timelineSteps);
-        return timelineSteps;
-      },
   
       /**
        * `closeModal`:
@@ -1528,6 +670,15 @@ import jsPDF from 'jspdf'
       },
   
       /**
+      openGuiaEmpresaModal(item) {
+        this.guiaEmpresaModalData = item
+        this.showGuiaEmpresaModal = true
+      },
+
+      closeGuiaEmpresaModal() {
+        this.guiaEmpresaModalData = null
+        this.showGuiaEmpresaModal = false
+      },
        * `getStatusChipClassTextual`:
        * Devuelve las clases CSS de Vuetify para aplicar un color específico
        * a los `v-chip` de estado, tanto para órdenes como para guías.
